add animation of skeleton
- fix bug with dmg of enemy
- fix bug with animation of dead (enemy)
- add to save component skeleton paramaters

+++ Assets/C#Scripts/Enemy/Moushroom/Spore.cs
@@ -11,7 +11,7 @@ public class Spore : MonoBehaviour
 
     private BoxCollider2D boxCollider; // оллайдер удара
 
-    private int sporeDamage = 20;
+    private float sporeDamage = 20;
     private float sporeCooldownDmg;
     GameObject player; //геймобьект игрок и ниже будет метод как он определ€етс€ и присваиваетс€ этой переменной
